# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export type MemberRole = typeof MemberRole[keyof typeof MemberRole];


export const MemberRole = {
  member: 'member',
  moderator: 'moderator',
  admin: 'admin',
  owner: 'owner',
} as const;

export interface CurrentAuthUser {
  user: AuthUser | null;
  isOwner: boolean;
  role: MemberRole;
}

export type AuthProvider = typeof AuthProvider[keyof typeof AuthProvider];


export const AuthProvider = {
  replit: 'replit',
  google: 'google',
  apple: 'apple',
} as const;

export interface AuthProviderOption {
  id: AuthProvider;
  label: string;
  enabled: boolean;
}

export interface AuthProviders {
  providers: AuthProviderOption[];
}

export interface Badge {
  slug: string;
  name: string;
  description: string;
  isDefault: boolean;
  isSelectable: boolean;
}

export interface MemberPublicProfile {
  userId: string;
  displayName: string;
  badge: Badge;
  createdAt: string;
}

export type MemberProfile = MemberPublicProfile & {
  role: MemberRole;
  earnedBadges: string[];
};

export type MemberProfileInputBadgeSlug = typeof MemberProfileInputBadgeSlug[keyof typeof MemberProfileInputBadgeSlug];


export const MemberProfileInputBadgeSlug = {
  'lok-clone': 'lok-clone',
  'lok-knight': 'lok-knight',
  'lokness-monster': 'lokness-monster',
  'lok-clown': 'lok-clown',
  'lok-loner': 'lok-loner',
  'lok-lit': 'lok-lit',
} as const;

export interface MemberProfileInput {
  /**
     * @minLength 2
     * @maxLength 24
     * @pattern ^[A-Za-z0-9][A-Za-z0-9 _-]{1,23}$
     */
  displayName?: string;
  badgeSlug?: MemberProfileInputBadgeSlug;
}

export interface MemberRoleRecord {
  userId: string;
  displayName: string;
  badge: Badge;
  role: MemberRole;
}

export type MemberRoleInputRole = typeof MemberRoleInputRole[keyof typeof MemberRoleInputRole];


export const MemberRoleInputRole = {
  member: 'member',
  moderator: 'moderator',
  admin: 'admin',
} as const;

export interface MemberRoleInput {
  role: MemberRoleInputRole;
}

export interface GameSocialStats {
  gameSlug: string;
  /** @minimum 0 */
  starCount: number;
  starred: boolean;
}

export interface GameStarInput {
  /**
     * @minLength 16
     * @maxLength 120
     */
  voterId: string;
  starred: boolean;
}

export interface GameFeedbackInput {
  /**
     * @minLength 4
     * @maxLength 2000
     */
  content: string;
}

export interface FeedbackReceipt {
  received: boolean;
  message: string;
}

export interface FeedbackAuthor {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
}

export type FeedbackNoteStatus = typeof FeedbackNoteStatus[keyof typeof FeedbackNoteStatus];


export const FeedbackNoteStatus = {
  pending: 'pending',
  reviewed: 'reviewed',
} as const;

export interface FeedbackNote {
  id: number;
  gameSlug: string;
  content: string;
  status: FeedbackNoteStatus;
  createdAt: string;
  updatedAt: string;
  author: FeedbackAuthor;
}

export interface FeedbackGroup {
  gameSlug: string;
  notes: FeedbackNote[];
}

export type FeedbackReviewInputStatus = typeof FeedbackReviewInputStatus[keyof typeof FeedbackReviewInputStatus];


export const FeedbackReviewInputStatus = {
  reviewed: 'reviewed',
} as const;

export interface FeedbackReviewInput {
  status: FeedbackReviewInputStatus;
}

export interface Error {
  error: string;
}

export type GetGameSocialStatsParams = {
/**
 * @maxLength 120
 */
voterId?: string;
};