# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /members/me:
    get:
      operationId: getMyMemberProfile
      tags: [members]
      summary: Get the signed-in member profile
      responses:
        "200":
          description: Current member profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberProfile"
        "401":
          description: Sign-in required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: updateMyMemberProfile
      tags: [members]
      summary: Update the signed-in member profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberProfileInput"
      responses:
        "200":
          description: Updated member profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberProfile"
        "400":
          description: Invalid profile update
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Sign-in required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Display name already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /members/{userId}:
    get:
      operationId: getMemberPublicProfile
      tags: [members]
      summary: Get a public member profile
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 255
      responses:
        "200":
          description: Public member profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberPublicProfile"
        "404":
          description: Member not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /members/roles:
    get:
      operationId: getMemberRoles
      tags: [members]
      summary: List member roles for the owner
      responses:
        "200":
          description: Member role assignments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MemberRoleRecord"
        "401":
          description: Sign-in required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Owner access required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /members/roles/{userId}:
    patch:
      operationId: updateMemberRole
      tags: [members]
      summary: Assign or revoke an admin or moderator role
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 255
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberRoleInput"
      responses:
        "200":
          description: Updated role assignment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberRoleRecord"
        "400":
          description: Invalid role update
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Sign-in required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Owner access required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Member not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
...
    Badge:
      type: object
      properties:
        slug:
          type: string
        name:
          type: string
        description:
          type: string
        isDefault:
          type: boolean
        isSelectable:
          type: boolean
      required: [slug, name, description, isDefault, isSelectable]
    MemberRole:
      type: string
      enum: [member, moderator, admin, owner]
    MemberPublicProfile:
      type: object
      properties:
        userId:
          type: string
        displayName:
          type: string
        badge:
          $ref: "#/components/schemas/Badge"
        createdAt:
          type: string
          format: date-time
      required: [userId, displayName, badge, createdAt]
    MemberProfile:
      allOf:
        - $ref: "#/components/schemas/MemberPublicProfile"
        - type: object
          properties:
            role:
              $ref: "#/components/schemas/MemberRole"
            earnedBadges:
              type: array
              items:
                type: string
          required: [role, earnedBadges]
    MemberProfileInput:
      type: object
      properties:
        displayName:
          type: string
          minLength: 2
          maxLength: 24
          pattern: "^[A-Za-z0-9][A-Za-z0-9 _-]{1,23}$"
        badgeSlug:
          type: string
          enum: [lok-clone, lok-knight, lokness-monster, lok-clown, lok-loner, lok-lit]
      minProperties: 1
      additionalProperties: false
    MemberRoleRecord:
      type: object
      properties:
        userId:
          type: string
        displayName:
          type: string
        badge:
          $ref: "#/components/schemas/Badge"
        role:
          $ref: "#/components/schemas/MemberRole"
      required: [userId, displayName, badge, role]
    MemberRoleInput:
      type: object
      properties:
        role:
          type: string
          enum: [member, moderator, admin]
      required: [role]
      additionalProperties: false
...
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]