# lib/api-zod/src/generated/types/badge.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Badge {
  slug: string;
  name: string;
  description: string;
  isDefault: boolean;
  isSelectable: boolean;
}