# lib/api-zod/src/generated/types/memberProfile.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MemberPublicProfile } from './memberPublicProfile';
import type { MemberRole } from './memberRole';

export type MemberProfile = MemberPublicProfile & {
  role: MemberRole;
  earnedBadges: string[];
};