# lib/api-zod/src/generated/types/memberProfileInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MemberProfileInputBadgeSlug } from './memberProfileInputBadgeSlug';

export interface MemberProfileInput {
  /**
     * @minLength 2
     * @maxLength 24
     * @pattern ^[A-Za-z0-9][A-Za-z0-9 _-]{1,23}$
     */
  displayName?: string;
  badgeSlug?: MemberProfileInputBadgeSlug;
}