# lib/api-zod/src/generated/types/memberPublicProfile.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Badge } from './badge';

export interface MemberPublicProfile {
  userId: string;
  displayName: string;
  badge: Badge;
  createdAt: Date;
}