# lib/api-zod/src/generated/types/memberRole.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type MemberRole = typeof MemberRole[keyof typeof MemberRole];


export const MemberRole = {
  member: 'member',
  moderator: 'moderator',
  admin: 'admin',
  owner: 'owner',
} as const;