# lib/api-zod/src/generated/types/memberRoleRecord.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Badge } from './badge';
import type { MemberRole } from './memberRole';

export interface MemberRoleRecord {
  userId: string;
  displayName: string;
  badge: Badge;
  role: MemberRole;
}